# modules/consumers/app/udaconnect/models/person.py
from sqlalchemy import Column, Integer, String
from app.udaconnect.infra.database import Base


class Person(Base):
    __tablename__ = "person"

    id = Column(Integer, primary_key=True)
    first_name = Column(String, nullable=False)
    last_name = Column(String, nullable=False)
    company_name = Column(String, nullable=False)